# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Namma Kasa API",
    "version": "0.1.0",
    "description": "Live tracking of door-to-door waste collection. Errors are RFC 9457 problem+json."
  },
  "servers": [
    {
      "url": "http://localhost:4000/v1",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/admin/driver-issues/wards/{wardId}": {
      "get": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Problems drivers have reported in this ward",
        "parameters": [
          {
            "in": "path",
            "name": "wardId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Issues",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriverIssueRecord"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/driver-issues/{id}/acknowledge": {
      "patch": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a reported issue as seen",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverIssueRecord"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/driver/issues": {
      "post": {
        "tags": [
          "driver"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Report a breakdown or blocked road to the Ward Admin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": [
                      "breakdown",
                      "road_blocked",
                      "other"
                    ]
                  },
                  "note": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "geo": {
                    "$ref": "#/components/schemas/GeoPoint"
                  }
                },
                "required": [
                  "kind"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverIssueRecord"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/admin/advisories": {
      "post": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Broadcast a schedule change to a route's residents",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "routeId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "note": {
                    "type": "string"
                  }
                },
                "required": [
                  "routeId",
                  "note"
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notified": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "notified"
                  ]
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/notifications/devices": {
      "post": {
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register this device for push",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fcmToken": {
                    "type": "string"
                  }
                },
                "required": [
                  "fcmToken"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Registered"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "default": "about:blank"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
      "GeoPoint": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "lng": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          }
        },
        "required": [
          "lat",
          "lng"
        ]
      },
...
      "DriverIssueRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "breakdown",
              "road_blocked",
              "other"
            ]
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "routeId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "driverName": {
            "type": "string"
          },
          "acknowledgedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "note",
          "routeId",
          "acknowledgedAt",
          "createdAt"
        ]
      }
    },
    "responses": {
      "Problem": {
        "description": "RFC 9457 problem document",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    }
  }
}